+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.20.0
 ### Breaking
 * replace print with display (more like scheme)
+* replace lips.Symbol with LSymbol
 ### Features
 * improve `-->` macro to handle numbers and strings and symbols (quoted or not)
 * new function `quoted-symbol?`